clean code
include header/footer/link
