+++ category-01.php
@@ -1,4 +1,4 @@
-<?php require'action.php'; 
+<?php require 'action.php'; 
 
 
 
@@ -10,235 +10,13 @@
 
 <head>
 	<title>category list</title>
-	<meta charset="UTF-8">
-	<meta name="viewport" content="width=device-width, initial-scale=1">
-	<!--===============================================================================================-->
-	<link rel="icon" type="image/png" href="images/icons/justice.png" />
-	<!--===============================================================================================-->
-	<link rel="stylesheet" type="text/css" href="vendor/bootstrap/css/bootstrap.min.css">
-	<link rel="stylesheet" type="text/css" href="vendor/bootstrap/css/bootstrap-rtl.min.css">
-	<!--===============================================================================================-->
-	<link rel="stylesheet" type="text/css" href="fonts/font-awesome-4.7.0/css/font-awesome.min.css">
-	<link rel="stylesheet" type="text/css" href="fonts/fontawesome-5.0.8/css/fontawesome-all.min.css">
-	<link href="https://fonts.googleapis.com/css2?family=Cairo&display=swap" rel="stylesheet">
-	<link href="https://fonts.googleapis.com/css2?family=Amiri&display=swap" rel="stylesheet">
-	<!--===============================================================================================-->
-	<link rel="stylesheet" type="text/css" href="fonts/iconic/css/material-design-iconic-font.min.css">
-	<!--===============================================================================================-->
-	<link rel="stylesheet" type="text/css" href="vendor/animate/animate.css">
-	<!--===============================================================================================-->
-	<link rel="stylesheet" type="text/css" href="vendor/css-hamburgers/hamburgers.min.css">
-	<!--===============================================================================================-->
-	<link rel="stylesheet" type="text/css" href="vendor/animsition/css/animsition.min.css">
-	<!--===============================================================================================-->
-	<link rel="stylesheet" type="text/css" href="css/util.min.css">
-	<!--===============================================================================================-->
-	<link rel="stylesheet" type="text/css" href="css/main.css">
-	<!--===============================================================================================-->
+	<?php include 'link.php'; ?>
 </head>
 
 <body class="animsition">
 
 	<!-- Header -->
-	<header>
-		<!-- Header desktop -->
-		<div class="container-menu-desktop">
-			<div class="topbar">
-				<div class="content-topbar container h-100">
-					<div class="left-topbar">
-						
-						<a href="login.php" class="left-topbar-item">
-							الدخول
-						</a>
-						
-						<a href="contact.php" class="left-topbar-item">
-							إتصل بنا
-						</a>
-						<a href="droit-pub.php" class="left-topbar-item">
-							سياسة النشر 
-						</a>
-
-						
-					</div>
-
-					<div class="right-topbar">
-						<a href="https://www.facebook.com/OCIAEJ">
-							<span class="fab fa-facebook-f"></span>
-						</a>
-
-						<a href="#">
-							<span class="fab fa-twitter"></span>
-						</a>
-
-						<a href="https://www.youtube.com/channel/UChBRIuAYSgkGaaxioE4U9jw">
-							<span class="fab fa-youtube"></span>
-						</a>
-					</div>
-				</div>
-			</div>
-
-			<!-- Header Mobile -->
-			<div class="wrap-header-mobile">
-				<!-- Button show menu -->
-				<div class="btn-show-menu-mobile hamburger hamburger--squeeze m-r--8">
-					<span class="hamburger-box">
-						<span class="hamburger-inner"></span>
-					</span>
-				</div>
-
-				<!-- Logo moblie -->		
-				<div class="logo-mobile">
-					<a href="index.php"><img src="images/logof.png" alt="IMG-LOGO"></a>
-				</div>
-	
-			</div>
-
-			<!-- Menu Mobile -->
-			<div class="menu-mobile">
-				<ul class="topbar-mobile">
-					
-
-					<li class="left-topbar">
-
-						<a href="login.php" class="left-topbar-item">
-							الدخول
-						</a>
-						<a href="contact.php" class="left-topbar-item">
-							إتصل بنا
-						</a>
-						<a href="droit-pub.php" class="left-topbar-item">
-							سياسة النشر 
-						</a>
-					</li>
-
-					<li class="right-topbar">
-						<a href="https://www.facebook.com/OCIAEJ">
-							<span class="fab fa-facebook-f"></span>
-						</a>
-
-						<a href="#">
-							<span class="fab fa-twitter"></span>
-						</a>
-
-						<a href="https://www.youtube.com/channel/UChBRIuAYSgkGaaxioE4U9jw">
-							<span class="fab fa-youtube"></span>
-						</a>
-					</li>
-				</ul>
-
-				<ul class="main-menu-m">
-					<li>
-						<a href="index.php">الرئيسية </a>
-
-						<span class="arrow-main-menu-m">
-							<i class="fa fa-angle-right" aria-hidden="true"></i>
-						</span>
-					</li>
-
-					<li>
-					<a href="blog-list-01.php?cat=فضاء التشريع">
-					    	فضاء التشريع </a>
-					</li>
-
-					<li>
-						<a href="blog-list-01.php?cat=مقالات قانونية">
-							مقالات قانونية </a>
-					</li>
-
-					<li>
-						<a hhref="blog-list-01.php?cat=إجتهاد قضائي">
-							إجتهاد قضائي </a>
-					</li>
-
-					<li>
-						<a href="blog-list-01.php?cat=أنشطة علمية">
-							أنشطة علمية </a>
-					</li>
-
-					<li>
-						<a href="blog-grid.php">صوت و صورة </a>
-					</li>
-
-				</ul>
-			</div>
-			
-			<!--  -->
-			<div class="wrap-logo container">
-				<!-- Logo desktop -->		
-				<div class="logo">
-					<a href="index.php"><img src="images/logo.jpg" alt="LOGO"></a>
-				</div>	
-
-				<!-- Banner -->
-				<div class="banner-header">
-					<a href="https://www.youtube.com/channel/UChBRIuAYSgkGaaxioE4U9jw"><img src="images/banner.png" alt="IMG"></a>
-				</div>
-			</div>	
-			
-			<!--  -->
-			<div class="wrap-main-nav">
-				<div class="main-nav">
-					<!-- Menu desktop -->
-					<nav class="menu-desktop">
-						<a class="logo-stick" href="index.php">
-							<img src="images/logof.png" alt="LOGO">
-						</a>
-
-						<ul class="main-menu">
-							<li class="main-menu-active">
-								<a href="index.php">الرئيسية </a>
-							</li>
-							<li>
-								<a href="blog-list-01.php?cat=فضاء التشريع">
-									فضاء التشريع </a>
-								<ul class="sub-menu">
-									<li><a href="category-01.php?sub=قوانين">
-									قوانين </a></li>
-									<li><a href="category-01.php?sub=مشاريع القوانين">
-									مشاريع القوانين </a></li>
-									<li><a href="category-01.php?sub=إتفاقيات">
-									إتفاقيات </a></li>
-									
-								</ul>
-							</li>
-							<li>
-								<a href="blog-list-01.php?cat=مقالات قانونية">
-									مقالات قانونية </a>
-								<ul class="sub-menu">
-									<li><a href="category-01.php?sub=دراسات">
-									دراسات  </a></li>
-									<li><a href="category-01.php?sub=وجهة نظر">
-									وجهة نظر  </a></li>
-									<li><a href="category-01.php?sub=تعليق على قرار">
-									تعليق على قرار  </a></li>
-									
-								</ul>
-							</li>
-							<li>
-								<a href="blog-list-01.php?cat=إجتهاد قضائي">
-									إجتهاد قضائي</a>
-							</li>
-							<li>
-								<a href="blog-list-01.php?cat=أنشطة علمية">
-									أنشطة علمية</a>
-								<ul class="sub-menu">
-									<li><a href="category-01.php?sub=فعاليات">
-									فعاليات  </a></li>
-									<li><a href="category-01.php?sub=تقارير">
-									تقارير </a></li>
-									
-								</ul>
-							</li>
-							<li>
-								<a href="blog-grid.php">صوت و صورة  </a>
-							</li>
-
-						</ul>
-					</nav>
-				</div>
-			</div>	
-		</div>
-	</header>
+	<?php include 'header.php'; ?>
 
 	<!-- Breadcrumb -->
 	<div class="container">
@@ -476,161 +254,7 @@
 	</section>
 
 	<!-- Footer -->
-	<footer>
-		<div class="bg2 p-t-40 p-b-25">
-			<div class="container">
-				<div class="row">
-					<div class="col-lg-4 p-b-20">
-						<div class="size-h-3 flex-s-c">
-							<h5 class="f1-m-7 cl0">
-								فضــــاء العدالة
-							</h5>
-						</div>
-
-						<div>
-							<p class="f1-s-1 cl11 p-b-16">
-								تهتم المجلة بنشر مختلف المواد التي لها علاقة بمجال القانون المغربي ، سواء المقالات القانونية أو الرسائل والأطروحات بما في ذلك الاجتهادات القضائية. وفي إطار تقريب المعلومة القانونية دائما، يقوم الموقع بدور الوسيط ولاسيما ما يخص النصوص القانونية المغربية والمقارنة، وكدا مختلف المجالات الاليكترونية والتي تهتم بالمجال القانون. وفي هذا الإطار لا يتحمل الموقع أية مسؤولية في المواد المنشور، وإنما ينسب الفكر أو التوجه لصاحب المقال، ونضع رهن إشارة الزائر، في حالة أي إخلال أو مس بالأمانة العلمية أو في ما يتعلق بأي محتوى خارج المادة العلمية المنشورة، التواصل معنا من أجل معالجة المشكل وذلك عبر البريد الإلكتروني
-							</p>
-
-							<p class="f1-s-5 cl11 p-b-16">
-								fadaa.al.adala@gmail.com
-							</p>
-
-							<div class="p-t-15">
-								<a href="https://www.facebook.com/OCIAEJ" class="fs-18 cl11 hov-cl10 trans-03 m-r-8">
-									<span class="fab fa-facebook-f"></span>
-								</a>
-
-								<a href="#" class="fs-18 cl11 hov-cl10 trans-03 m-r-8">
-									<span class="fab fa-twitter"></span>
-								</a>
-
-								<a href="https://www.youtube.com/channel/UChBRIuAYSgkGaaxioE4U9jw" class="fs-18 cl11 hov-cl10 trans-03 m-r-8">
-									<span class="fab fa-youtube"></span>
-								</a>
-							</div>
-						</div>
-					</div>
-
-					<div class="col-sm-6 col-lg-4 p-b-20">
-						<div class="size-h-3 flex-s-c">
-							<h5 class="f1-m-7 cl0">
-								 الأكثر متـــابعة
-							</h5>
-						</div>
-
-						<ul>
-							<li class="flex-wr-sb-s p-b-20">
-								<a href="#" class="size-w-4 wrap-pic-w hov1 trans-03">
-									<img src="images/popular-post-01.jpg" alt="IMG">
-								</a>
-
-								<div class="size-w-5">
-									<h6 class="p-b-5">
-										<a href="#" class="f1-s-5 cl11 hov-cl10 trans-03">
-											جائحة كورونا تجمع أساتذة قانون من المغرب والجزائر والعراق في ندوة افتراضية
-										</a>
-									</h6>
-
-									<span class="f1-s-3 cl6">
-										22 فبراير 2020
-									</span>
-								</div>
-							</li>
-
-							<li class="flex-wr-sb-s p-b-20">
-								<a href="#" class="size-w-4 wrap-pic-w hov1 trans-03">
-									<img src="images/popular-post-02.jpg" alt="IMG">
-								</a>
-
-								<div class="size-w-5">
-									<h6 class="p-b-5">
-										<a href="#" class="f1-s-5 cl11 hov-cl10 trans-03">
-											إشكالية التمييز بين خبرة التسيير والخبرة المحاسبية
-										</a>
-									</h6>
-
-									<span class="f1-s-3 cl6">
-										22 فبراير 2020
-									</span>
-								</div>
-							</li>
-
-							<li class="flex-wr-sb-s p-b-20">
-								<a href="#" class="size-w-4 wrap-pic-w hov1 trans-03">
-									<img src="images/popular-post-03.jpg" alt="IMG">
-								</a>
-
-								<div class="size-w-5">
-									<h6 class="p-b-5">
-										<a href="#" class="f1-s-5 cl11 hov-cl10 trans-03">
-											ورقة حول موضوع الشرعية الجنائية الإجرائية في زمن كورونا
-										</a>
-									</h6>
-
-									<span class="f1-s-3 cl6">
-										22 فبراير 2020
-									</span>
-								</div>
-							</li>
-						</ul>
-					</div>
-
-					<div class="col-sm-6 col-lg-4 p-b-20">
-						<div class="size-h-3 flex-s-c">
-							<h5 class="f1-m-7 cl0">
-								الأقســــام 
-							</h5>
-						</div>
-
-						<ul class="m-t--12">
-							<li class="how-bor1 p-rl-5 p-tb-10">
-								<a href="blog-list-01.php" class="f1-s-5 cl11 hov-cl10 trans-03 p-tb-8">
-									فضاء التشريع (22)
-								</a>
-							</li>
-
-							<li class="how-bor1 p-rl-5 p-tb-10">
-								<a href="blog-list-01.php" class="f1-s-5 cl11 hov-cl10 trans-03 p-tb-8">
-									مقالات قانونية  (22)
-								</a>
-							</li>
-
-							<li class="how-bor1 p-rl-5 p-tb-10">
-								<a href="blog-list-01.php" class="f1-s-5 cl11 hov-cl10 trans-03 p-tb-8">
-									إجتهاد قضائي (22)
-								</a>
-							</li>
-
-							<li class="how-bor1 p-rl-5 p-tb-10">
-								<a href="blog-list-01.php" class="f1-s-5 cl11 hov-cl10 trans-03 p-tb-8">
-								    أنشطة علمية (22)
-								</a>
-							</li>
-
-							<li class="how-bor1 p-rl-5 p-tb-10">
-								<a href="blog-grid.php" class="f1-s-5 cl11 hov-cl10 trans-03 p-tb-8">
-								    صوت و صورة (22)
-								</a>
-							</li>
-						</ul>
-					</div>
-				</div>
-			</div>
-		</div>
-
-		<div class="bg11">
-			<div class="container size-h-4 flex-c-c p-tb-15">
-				<span class="f1-s-1 cl0 txt-center">
-					Copyright © 2018 
-
-					<a href="#" class="f1-s-1 cl10 hov-link1"><!-- Link back to Colorlib can't be removed. Template is licensed under CC BY 3.0. -->
-Copyright &copy;<script>document.write(new Date().getFullYear());</script> All rights reserved | This template is made with <i class="fa fa-heart" aria-hidden="true"></i> by <a href="https://colorlib.com" target="_blank">Colorlib</a>
-<!-- Link back to Colorlib can't be removed. Template is licensed under CC BY 3.0. -->
-				</span>
-			</div>
-		</div>
-	</footer>
+	<?php include'footer.php'; ?>
 
 	<!-- Back to top -->
 	<div class="btn-back-to-top" id="myBtn">
