+++ dashboard.php
@@ -1,4 +1,4 @@
-<?php require'action.php'; 
+<?php require 'action.php'; 
 
 /*
 <?php
