+++ header.php
@@ -1,0 +1,199 @@
+<header>
+		<!-- Header desktop -->
+		<div class="container-menu-desktop">
+			<div class="topbar">
+				<div class="content-topbar container h-100">
+					<div class="left-topbar">
+						
+						<a href="login.php" class="left-topbar-item">
+							الدخول
+						</a>
+						
+						<a href="contact.php" class="left-topbar-item">
+							إتصل بنا
+						</a>
+						<a href="droit-pub.php" class="left-topbar-item">
+							سياسة النشر 
+						</a>
+
+						
+					</div>
+
+					<div class="right-topbar">
+						<a href="https://www.facebook.com/OCIAEJ">
+							<span class="fab fa-facebook-f"></span>
+						</a>
+
+						<a href="#">
+							<span class="fab fa-twitter"></span>
+						</a>
+
+						<a href="https://www.youtube.com/channel/UChBRIuAYSgkGaaxioE4U9jw">
+							<span class="fab fa-youtube"></span>
+						</a>
+					</div>
+				</div>
+			</div>
+
+			<!-- Header Mobile -->
+			<div class="wrap-header-mobile">
+				<!-- Button show menu -->
+				<div class="btn-show-menu-mobile hamburger hamburger--squeeze m-r--8">
+					<span class="hamburger-box">
+						<span class="hamburger-inner"></span>
+					</span>
+				</div>
+
+				<!-- Logo moblie -->		
+				<div class="logo-mobile">
+					<a href="index.php"><img src="images/logof.png" alt="IMG-LOGO"></a>
+				</div>
+	
+			</div>
+
+			<!-- Menu Mobile -->
+			<div class="menu-mobile">
+				<ul class="topbar-mobile">
+					
+
+					<li class="left-topbar">
+
+						<a href="login.php" class="left-topbar-item">
+							الدخول
+						</a>
+						<a href="contact.php" class="left-topbar-item">
+							إتصل بنا
+						</a>
+						<a href="droit-pub.php" class="left-topbar-item">
+							سياسة النشر 
+						</a>
+					</li>
+
+					<li class="right-topbar">
+						<a href="https://www.facebook.com/OCIAEJ">
+							<span class="fab fa-facebook-f"></span>
+						</a>
+
+						<a href="#">
+							<span class="fab fa-twitter"></span>
+						</a>
+
+						<a href="https://www.youtube.com/channel/UChBRIuAYSgkGaaxioE4U9jw">
+							<span class="fab fa-youtube"></span>
+						</a>
+					</li>
+				</ul>
+
+				<ul class="main-menu-m">
+					<li>
+						<a href="index.php">الرئيسية </a>
+
+						<span class="arrow-main-menu-m">
+							<i class="fa fa-angle-right" aria-hidden="true"></i>
+						</span>
+					</li>
+
+					<li>
+					<a href="blog-list-01.php?cat=فضاء التشريع">
+					    	فضاء التشريع </a>
+					</li>
+
+					<li>
+						<a href="blog-list-01.php?cat=مقالات قانونية">
+							مقالات قانونية </a>
+					</li>
+
+					<li>
+						<a hhref="blog-list-01.php?cat=إجتهاد قضائي">
+							إجتهاد قضائي </a>
+					</li>
+
+					<li>
+						<a href="blog-list-01.php?cat=أنشطة علمية">
+							أنشطة علمية </a>
+					</li>
+
+					<li>
+						<a href="blog-grid.php">صوت و صورة </a>
+					</li>
+
+				</ul>
+			</div>
+			
+			<!--  -->
+			<div class="wrap-logo container">
+				<!-- Logo desktop -->		
+				<div class="logo">
+					<a href="index.php"><img src="images/logo.jpg" alt="LOGO"></a>
+				</div>	
+
+				<!-- Banner -->
+				<div class="banner-header">
+					<a href="https://www.youtube.com/channel/UChBRIuAYSgkGaaxioE4U9jw"><img src="images/banner.png" alt="IMG"></a>
+				</div>
+			</div>	
+			
+			<!--  -->
+			<div class="wrap-main-nav">
+				<div class="main-nav">
+					<!-- Menu desktop -->
+					<nav class="menu-desktop">
+						<a class="logo-stick" href="index.php">
+							<img src="images/logof.png" alt="LOGO">
+						</a>
+
+						<ul class="main-menu">
+							<li class="main-menu-active">
+								<a href="index.php">الرئيسية </a>
+							</li>
+							<li>
+								<a href="blog-list-01.php?cat=فضاء التشريع">
+									فضاء التشريع </a>
+								<ul class="sub-menu">
+									<li><a href="category-01.php?sub=قوانين">
+									قوانين </a></li>
+									<li><a href="category-01.php?sub=مشاريع القوانين">
+									مشاريع القوانين </a></li>
+									<li><a href="category-01.php?sub=إتفاقيات">
+									إتفاقيات </a></li>
+									
+								</ul>
+							</li>
+							<li>
+								<a href="blog-list-01.php?cat=مقالات قانونية">
+									مقالات قانونية </a>
+								<ul class="sub-menu">
+									<li><a href="category-01.php?sub=دراسات">
+									دراسات  </a></li>
+									<li><a href="category-01.php?sub=وجهة نظر">
+									وجهة نظر  </a></li>
+									<li><a href="category-01.php?sub=تعليق على قرار">
+									تعليق على قرار  </a></li>
+									
+								</ul>
+							</li>
+							<li>
+								<a href="blog-list-01.php?cat=إجتهاد قضائي">
+									إجتهاد قضائي</a>
+							</li>
+							<li>
+								<a href="blog-list-01.php?cat=أنشطة علمية">
+									أنشطة علمية</a>
+								<ul class="sub-menu">
+									<li><a href="category-01.php?sub=فعاليات">
+									فعاليات  </a></li>
+									<li><a href="category-01.php?sub=تقارير">
+									تقارير </a></li>
+									
+								</ul>
+							</li>
+							<li>
+								<a href="blog-grid.php">صوت و صورة  </a>
+							</li>
+
+						</ul>
+					</nav>
+				</div>
+			</div>	
+		</div>
+	</header>
